feat(pos): change dineInPrice type to BigDecimal and update order item pricing logic

+++ src/main/java/com/cafe/erp/modules/catalogue/item/dto/ItemRequestDTO.java
@@ -25,7 +25,7 @@ public class ItemRequestDTO {
 
     // Pricing
     private BigDecimal basePrice;
-    private Double dineInPrice;
+    private BigDecimal dineInPrice;
     private Double takeawayPrice;
     private Double deliveryPrice;
     private Double aggregatorPrice;
